chore: Update Course.js to use lowercase input for hashing
This commit updates the Course.js file to use lowercase input for hashing instead of using the SHA256 algorithm. The previous code was using CryptoJS.SHA256(input.toLowerCase()) to hash the input, but it has been changed to simply input.toLowerCase(). This change improves the code by removing unnecessary hashing and simplifying the logic.